clients/py/pyproject.toml
[tool.poetry]
name = "py"
version = "0.1.0"
description = ""
authors = ["DBCodes <dennis@dbproductions.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pika = "^1.3.2"
rstream = "^0.15.0"
paho-mqtt = "^1.6.1"
stomp-py = "^8.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
